==> Data_Jobs_Terzic_Project/0_Data_cleaning_and_merging.ipynb <==
## Data Cleaning and Merging

This section covers the steps taken to prepare the raw datasets for analysis:
- Load Raw Data
- Clean Data
- Merge Data
- Save Final Dataset

### Import Libraries

from pathlib import Path
import ast

import requests
import pandas as pd

### Load Raw Data

#### Import the job postings dataset

# %pip install datasets
from datasets import load_dataset
dataset = load_dataset('lukebarousse/data_jobs')
df = dataset['train'].to_pandas()

df.head(3)

To make the salary data more relevant for European users, I extended the main dataset (which includes yearly salaries in USD) with two additional sources:
- **Exchange Rate** — to convert salaries from USD to EUR.
- **Country Dictionary** — to group EU countries for regional analysis.

#### Load USD→EUR exchange rate from a public API

def get_usd_to_eur_rate_frankfurter(year):
    date = f"{year}-01-01"
    url = f"https://api.frankfurter.app/{date}?from=USD&to=EUR"
    response = requests.get(url)
    data = response.json()
    return data['rates']['EUR'] if 'rates' in data and 'EUR' in data['rates'] else None

# Just 2023
years = [2023]

# Get exchange rate for 2023
ex_rates = []
for y in years:
    rate = get_usd_to_eur_rate_frankfurter(y)
    ex_rates.append({'year': y, 'usd_to_eur': rate})

df_ex_rate = pd.DataFrame(ex_rates).round(4)

raw_data_dir = Path.cwd().parents[1] / 'Raw_Data'
df_ex_rate.to_csv(raw_data_dir / 'ex_rate_2.csv', index=False)

df_ex_rate.head(3)

#### Load a countries dictionary

df_EU = pd.read_csv(raw_data_dir / 'EU_Countries_dict.csv', delimiter=';')

df_EU.head(3)

### Clean Data

df = df.dropna(subset=['salary_year_avg'])
df = df.drop_duplicates()
df['job_posted_date'] = pd.to_datetime(df['job_posted_date'])
df['job_skills'] = df['job_skills'].apply(lambda x: ast.literal_eval(x) if pd.notna(x) else x)

### Merge Data

#### Merge Exchange rate
In this section we are merging Exchange rate to the main dataset and convert annual salaries to EUR. 
Since salaries in Europe are commonly discussed monthly, we are calculating monthly equivalents for better local insight.

# Merge Exchange Rate. Convert job_year_avg in EUR
df_Final = df.rename(columns={'salary_year_avg': 'salary_year_avg_usd'})
df_Final['year'] = pd.to_datetime(df_Final['job_posted_date']).dt.year
df_Final = df_Final.merge(df_ex_rate, how='left', on='year')
df_Final['salary_year_avg_eur'] = (df_Final['salary_year_avg_usd'] * df_Final['usd_to_eur']).round(2)

# Calculate monthly salaries
df_Final['salary_month_avg_eur'] = (df_Final['salary_year_avg_eur'] / 12).round(2)

df_Final.loc[1:3, ['job_title_short', 'usd_to_eur', 'salary_year_avg_usd', 'salary_year_avg_eur', 'salary_month_avg_eur']]

#### Merge Countries dictionary

To analyze the European data jobs market add is_eu and ISO columns from the countries dictionary.
- Use is_eu column to group countries that are European Union members.
- Use ISO codes to build interactive map with job salaries in 3_Salary_Analysis notebook.

df_Final = df_Final.merge(df_EU, how='left', left_on='job_country', right_on='country')
pd.set_option('future.no_silent_downcasting', True)
df_Final['is_eu'] = df_Final['is_eu'].fillna(False).astype(bool)

df_Final.loc[1:10, ['job_title_short', 'job_country', 'is_eu', 'ISO']]

# Add Region grouping column
def group_country(row):
    if row['job_country'] == 'United States':
        return 'US'
    elif row['is_eu']:
        return 'EU'
    else:
        return 'Other'
    
df_Final['region_group'] = df_Final.apply(group_country, axis=1)

df_Final.loc[1:10, ['region_group', 'job_country', 'salary_month_avg_eur']]

### Save Final Dataset

df_Final.drop(columns=['usd_to_eur', 'year', 'salary_year_avg_usd', 'salary_hour_avg'], inplace=True)
df_Final.to_csv(raw_data_dir / 'df_Final.csv', index=False)

df_Final.head(3)
